scripts/rag_search/requirements.txt
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
numpy==1.24.3
google-generativeai==0.3.2
openai==1.6.1
python-dotenv==1.0.0
scikit-learn==1.3.2